+++ packages/ts-newskit/src/theme/index.ts
@@ -1,2 +1,3 @@
 // @ts-nocheck
 export { TimesWebLightTheme } from './times-web-light';
+export { PuzzlesWebLightTheme } from './puzzles-web-light';