+++ packages/ts-newskit/src/theme/times-web-light/style-presets/flag.ts
@@ -1,0 +1,18 @@
+export const flagStylePresets = {
+  flagProgress: {
+    base: {
+      backgroundColor: '{{colors.interfaceInformative010}}',
+      color: '{{colors.inkInverse}}',
+      borderRadius: '2px',
+      textTransform: 'uppercase'
+    }
+  },
+  flagComplete: {
+    base: {
+      backgroundColor: '{{colors.interfacePositive010}}',
+      color: '{{colors.inkInverse}}',
+      borderRadius: '2px',
+      textTransform: 'uppercase'
+    }
+  }
+};