+++ packages/ts-newskit/src/theme/times-web-light/style-presets/index.ts
@@ -9,7 +9,8 @@ import { pageStylePresets } from './page';
 import { secondaryNavigationStylePresets } from './secondary-nav';
 import { sliceStylePresets } from './slices';
 import { textStylePresets } from './text';
-import { breakoutSectionStylePresets } from './breakoutSection';
+import { flagStylePresets } from './flag';
+import { accordionStylePresets } from './accordion';
 
 export const stylePresets = {
   ...articleListStylePresets,
@@ -23,5 +24,6 @@ export const stylePresets = {
   ...secondaryNavigationStylePresets,
   ...leadStoryStylePresets,
   ...sliceStylePresets,
-  ...breakoutSectionStylePresets
+  ...flagStylePresets,
+  ...accordionStylePresets
 };