+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.72.1](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.72.0...@times-components/article-main-standard@3.72.1) (2023-04-19)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
 # [3.72.0](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.71.11...@times-components/article-main-standard@3.72.0) (2023-04-18)